+++ src/main/java/Main.java
@@ -9,9 +9,9 @@ import java.util.concurrent.ExecutionException;
 
 public class Main {
     public static void main(String[] args) throws ExecutionException, InterruptedException {
-        //Modelo modelo = new Modelo();
+        Vista vista = new NewsGUI();
         //Controlador controlador = new Controlador(new Modelo());
-        Usuario usuario1 = new Usuario("lOCO", new Controlador(new Modelo()));
+        Usuario usuario1 = new Usuario("lOCO", new Controlador(new Modelo(), vista));
 
         usuario1.buscar();
     }
